WIP - separating view, controller, and file processing service
